+++ packages/tolk/test/unit/uop/test_uop.ml
@@ -129,6 +129,7 @@ let ops_tinygrad_order () =
       "IF";
       "END";
       "ENDIF";
+      "BACKEDGE";
       "WAIT";
       "CONST";
       "CUSTOM";